backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23
python-multipart==0.0.6

# Optional: for PostgreSQL support
# psycopg2-binary==2.9.9

# Optional: for MySQL support  
# pymysql==1.1.0